The purpose of this commit is to align the interface specification with the actual implementation of the provisioning interface.
The version property was originally removed in june 2015 with commit 364856f3955344e4b511853eeca593851c7becdd, however it was never removed from the actual implementation of Jobfunktionsrolle (JFR) in the provisioning interface for CH. To ensure that the interface definition corresponds to the acual implementation, it has been re-introduced here.

The version property is not actively used, since the Version information is encoded in the Privilege value for a JFR, which always ends with the version number of the JFR, e.g. /1. As such this property is redundant, and it is not possible to support it using existing federation servers since there's no concept of "version" on an OIO BPP Privilige, and consequently no way to convey this information - so it has been encoded in the privilege name instead.